Modularized macro expansion (changed API)
This commit is a from-scratch rewrite of the library, changing
the way code is generated and breaking with the previous API.

The primary goal is being able to define custom Record converters,
which was previously impossible due to the way code was generated
(everything in one pass).

Incidentally, this rewrite also fixes a previous bug which made unable
to compile records with fields named after Elixir keywords (e.g.,
':fn').

Documentation not yet updated.